Binary Classification

We've implemented some basic functionality for binary classification tasks, including a simple perceptron model and logistic regression. This model can be trained on labeled data and used to classify new instances.

VLDataScienceMachineLearningPackage.learn — Function

learn(features::Array{<:Number,2}, labels::Array{<:Number,1}, algorithm::AbstractClassificationAlgorithm; 
    maxiter::Int64 = 100, verbose::Bool = false)

The function learns a classification model from the data provided using the algorithm specified. This is a wrapper function that calls the internal function _learn whose implementation is algorithm-specific.

Arguments

features::Array{<:Number,2}: the features.

labels::Array{<:Number,1}: the labels.

algorithm::AbstractClassificationAlgorithm: the algorithm to use to learn the model.

Returns

the updated algorithm model.

VLDataScienceMachineLearningPackage.classify — Function

classify(features::Array{<:Number,2}, algorithm::AbstractClassificationAlgorithm)

classify(test::Array{<:Number,1}, algorithm::AbstractClassificationAlgorithm)

VLDataScienceMachineLearningPackage.confusion — Function

confusion(actual::Array{Int64,1}, model::Array{Int64,1}) -> Array{Int64,2}

The function computes the confusion matrix for the classification model.

Arguments

actual::Array{<:Number,1}: the actual labels.

model::Array{<:Number,1}: the model estimated labels.

Returns

a 2x2 confusion matrix. The rows correspond to the actual labels and the columns correspond to the predicted labels.