Clustering and Classification Algorithms

This section provides an overview of various clustering algorithms implemented in the VLDataScienceMachineLearningPackage.jl. Clustering is an unsupervised learning technique used to group similar data points together based on their features.

K-Means Clustering

K-Means is a popular clustering algorithm that partitions data into K distinct clusters based on feature similarity. The algorithm iteratively assigns data points to the nearest cluster centroid and updates the centroids until convergence.

VLDataScienceMachineLearningPackage.MyNaiveKMeansClusteringAlgorithm — Type

mutable struct MyNaiveKMeansClusteringAlgorithm <: MyAbstractUnsupervisedClusteringAlgorithm

A mutable struct that represents a naive K-Means clustering algorithm.

Fields

K::Int64: number of clusters

centroids::Dict{Int64, Vector{Float64}}: cluster centroids

assignments::Vector{Int64}: cluster assignments

ϵ::Float64: convergence criteria

maxiter::Int64: maximum number of iterations

dimension::Int64: dimension of the data

number_of_points::Int64: number of data points

VLDataScienceMachineLearningPackage.cluster — Function

cluster(data::Array{<:Number,2}, algorithm::T; d = Euclidean(), verbose::Bool = false) where T <: MyAbstractUnsupervisedClusteringAlgorithm

Cluster the input data using the specified clustering algorithm.

Arguments

data::Array{<:Number,2}: A 2D array where each row represents a data point.

algorithm::T: An instance of a clustering algorithm type T that is a subtype of MyAbstractUnsupervisedClusteringAlgorithm.

d: (Optional) A distance metric function. Default is Euclidean().

verbose::Bool: (Optional) A boolean flag to enable verbose output. Default is false.

tmpdir::String: (Optional) A directory path for temporary files. Default is the current directory ".".

Returns

A tuple containing the clustering assignments, centroids, and the number of iterations taken to converge.