Data

We've included several datasets in the package that we use for examples, activities, etc.

VLDataScienceMachineLearningPackage.MyKaggleCustomerSpendingDataset — Function

MyKaggleCustomerSpendingDataset() -> DataFrame

Load the Kaggle customer spending dataset as a DataFrame. The original dataset can be found at: Spending dataset.

VLDataScienceMachineLearningPackage.MyStringDecodeChallengeDataset — Function

MyStringDecodeChallengeDataset() -> NamedTuple

Load the String Decode Challenge testing and production datasets.

Return

NamedTuple: A tuple containing the three datasets:

test_part_1: The first part of the test dataset.

test_part_2: The second part of the test dataset.

production: The production dataset.

VLDataScienceMachineLearningPackage.MyCommonSurnameDataset — Function

MyCommonSurnameDataset() -> DataFrame

Load the common surnames dataset by country as a DataFrame. The original dataset can be found at: Common Surnames by Country.

VLDataScienceMachineLearningPackage.MyCommonForenameDataset — Function

MyCommonForenameDataset() -> DataFrame

Load the common forenames dataset by country as a DataFrame. The original dataset can be found at: Common Forenames by Country.

VLDataScienceMachineLearningPackage.MyKaggleHousingPricesDataset — Function

MyHousingPricesDataset() -> DataFrame

Load the house prices dataset from Kaggle as a DataFrame. The original dataset can be found at: Housing Prices Dataset on Kaggle

VLDataScienceMachineLearningPackage.MySarcasmCorpus — Function

function MySarcasmCorpus() -> MySarcasmRecordCorpusModel

The function corpus reads a file composed of JSON records and returns the data as a MySarcasmRecordCorpusModel instance. Each record in the file is expected to have the following fields:

is_sarcastic::Bool - a boolean value indicating if the headline is sarcastic.

headline::String - the headline of the article.

article_link::String - the link to the article.

Returns

MySarcasmRecordCorpusModel - the data from the file as a MySarcasmRecordCorpusModel instance.

VLDataScienceMachineLearningPackage.MySMSSpamHamCorpus — Function

function MySMSSpamHamCorpus() -> MySMSSpamHamRecordCorpusModel

The function MySMSSpamHamCorpus reads the SMS Spam Ham dataset and returns the data as a MySMSSpamHamRecordCorpusModel instance.

VLDataScienceMachineLearningPackage.MyGraphEdgeModels — Function

function MyGraphEdgeModels(filepath::String, edgeparser::Function; comment::Char='#', 
delim::Char=',')::Dict{Int64,MyGraphEdgeModel}

Function to parse an edge file and return a dictionary of edges models.

Arguments

filepath::String: The path to the edge file.

edgeparser::Function: A callback function to parse each edge line. This function should take a line as input, and a delimiter character, and return a tuple of the form (source, target, data), where:

source::Int64: The source node ID.

target::Int64: The target node ID.

data::Any: Any additional data associated with the edge, e.g., a weight, a tuple of information, etc.

Returns

Dict{Int64,MyGraphEdgeModel}: A dictionary of edge models.

VLDataScienceMachineLearningPackage.MyConstrainedGraphEdgeModels — Function

function MyConstrainedGraphEdgeModels(filepath::String, edgeparser::Function; comment::Char='#', 
    delim::Char=',') -> Dict{Int64,MyConstrainedGraphEdgeModel}

This function parses a constrained graph edge file and returns a dictionary of constrained graph edge models.

Arguments

filepath::String: The path to the edge file.

edgeparser::Function: A callback function to parse each edge line. This function should take a line as input, and a delimiter character, and return a tuple of the form (source, target, weight, lower, upper), where:

source::Int64: The source node ID.

target::Int64: The target node ID.

weight::Union{Nothing, Number}: The weight of the edge.

lower::Union{Nothing, Number}: The lower bound of the edge weight.

upper::Union{Nothing, Number}: The upper bound of the edge weight.

Returns

Dict{Int64,MyConstrainedGraphEdgeModel}: A dictionary of constrained graph edge models.

VLDataScienceMachineLearningPackage.MyBanknoteAuthenticationDataset — Function

MyBanknoteAuthenticationDataset() -> DataFrame

The second dataset we will explore is the banknote authentication dataset from the UCI archive. This dataset has 1372 instances of 4 continuous features and an integer (-1,1) class variable.

Financial Data

We have included a dataset of market data that we use for examples and activities in the financial domain. This dataset holds the daily open, high, low, close, and volume data for a selection of stocks between 2014 and 2024. In addition, we have defined a few methods for working with this data.

VLDataScienceMachineLearningPackage.MyTrainingMarketDataSet — Function

MyTrainingMarketDataSet() -> Dict{String, DataFrame}

Load the components of the SP500 Daily open, high, low, close (OHLC) dataset as a dictionary of DataFrames. This data was provided by Polygon.io and covers the period from January 3, 2014, to December 31, 2024.

VLDataScienceMachineLearningPackage.log_growth_matrix — Function

function log_growth_matrix(dataset::Dict{String, DataFrame}, 
            firms::Array{String,1}; Δt::Float64 = (1.0/252.0), risk_free_rate::Float64 = 0.0) -> Array{Float64,2}

The log_growth_matrix function computes the excess log growth matrix for a given set of firms where we define the log growth as:

\[ \mu_{t,t-1}(r_{f}) = \frac{1}{\Delta t} \log\left(\frac{S_{t}}{S_{t-1}}\right) - r_f\]

where $S_t$ is the volume weighted average price (units: USD/share) at time t, $\Delta t$ is the time increment (in years), and $r_f$ is the annual risk-free rate (units: 1/years) assuming continuous compounding.

Arguments

dataset::Dict{String, DataFrame}: A dictionary of data frames where the keys are the firm ticker symbols and the values are the data frames holding price data. We use the volume_weighted_average_price column to compute the log growth by default.

firms::Array{String,1}: An array of firm ticker symbols for which we want to compute the log growth matrix.

Δt::Float64: The time increment used to compute the log growth. The default value is 1/252, i.e., one trading day in units of years.

risk_free_rate::Float64: The risk-free rate used to compute the log growth. The default value is 0.0.

keycol::Symbol: The column in the data frame to use to compute the log growth. The default value is :volume_weighted_average_price.

testfirm::String: The firm ticker symbol to use to determine the number of trading days. By default, we use "AAPL".

Returns

Array{Float64,2}: An array of the excess log growth values for the given set of firms. The time series is the rows and the firms are the columns. The columns are ordered according to the order of the firms array.

See:

The DataFrame type (and methods for working with data frames) is exported from the DataFrames.jl package