Factory methods

We use a particular pattern to build and configure the various composite types in our system. This pattern involves the use of factory functions that encapsulate the construction logic for each type, ensuring that all necessary dependencies are properly initialized and configured.

VLDataScienceMachineLearningPackage.build — Function

function build(type::Type{MyAdjacencyRecombiningCommodityPriceTree}, data::NamedTuple) -> MyAdjacencyRecombiningCommodityPriceTree

Builds an MyAdjacencyRecombiningCommodityPriceTree model given the data in the NamedTuple. This method builds the connectivity of the tree. To compute the price at each node, use the populate! method.

Arguments

type::Type{MyAdjacencyRecombiningCommodityPriceTree}: The type of the model to build.

data::NamedTuple: The data to use to build the model.

The data NamedTuple must contain the following fields:

h::Int64: The height of the tree.

price::Float64: The price at the root node.

u::Float64: The price increase factor.

d::Float64: The price decrease factor.

Returns

MyAdjacencyRecombiningCommodityPriceTree: the price tree model holding the computed price data.

function build(modeltype::Type{MyOneDimensionalElementaryWolframRuleModel}, data::NamedTuple) -> MyOneDimensionalElementarWolframRuleModel

This build method constructs an instance of the MyOneDimensionalElementaryWolframRuleModel type using the data in a NamedTuple.

Arguments

modeltype::Type{MyOneDimensionalElementaryWolframRuleModel}: The type of model to build, in this case, the MyOneDimensionalElementaryWolframRuleModel type.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

index::Int64: The index of the Wolfram rule

colors::Int64: The number of colors in the rule

radius::Int64: The radius, i.e., the number of cells to consider in the rule

Return

This function returns a populated instance of the MyOneDimensionalElementaryWolframRuleModel type.

VLDataScienceMachineLearningPackage.populate! — Function

populate!(model::MyAdjacencyRecombiningCommodityPriceTree, price::Float64, Δ::Array{Float64,1})

This function populates the price tree model with the given price and price change factors. This method updates the model in place.

Arguments

model::MyAdjacencyRecombiningCommodityPriceTree: The price tree model to populate.

price::Float64: The initial price to set at the root of the tree.

Δ::Array{Float64,1}: The array of price change factors for each level of the tree.