Multiplicative Weight Updates

This section describes the implementation of the Binary Weighted Majority Algorithm (WMA) and Multiplicative Weights Algorithm (MWA) for online learning in the VLDataScienceMachineLearningPackage.jl Julia package.

VLDataScienceMachineLearningPackage.MyBinaryWeightedMajorityAlgorithmModel — Type

MyBinaryWeightedMajorityAlgorithmModel

A mutable type for the Binary Weighted Majority Algorithm model. This model is used to simulate the Binary Weighted Majority Algorithm. The model has the following fields:

ϵ::Float64: learning rate

n::Int64: number of experts

T::Int64: number of rounds

weights::Array{Float64,2}: weights of the experts

expert::Function: expert function

adversary::Function: adversary function

VLDataScienceMachineLearningPackage.play — Function

play(model::MyBinaryWeightedMajorityAlgorithmModel, data::Array{Float64,2})

Play the Binary Weighted Majority Algorithm. This function simulates the Binary Weighted Majority Algorithm using the given model and data. The function returns a tuple with two elements. The first element is a matrix with the results of the simulation. The second element is the weights of the experts at the end of the simulation.

Arguments

model::MyBinaryWeightedMajorityAlgorithmModel: the model to simulate

world::Function: the world function that generates the actual outcomes

data::Array{Float64,2}: the data to use in the simulation

Returns

Tuple{Array{Int64,2}, Array{Float64,2}}: a tuple with the results of the simulation and the weights of the experts