Q-Learning

Q-Learning is a model-free reinforcement learning algorithm that aims to learn the value of taking a particular action in a given state. It does this by iteratively updating a Q-value table based on the agent's experiences in the environment.

VLDataScienceMachineLearningPackage.MyQLearningAgentModel — Type

mutable struct MyQLearningAgentModel <: AbstractOnlineLearningModel

A mutable type for the Q-Learning Agent model.

Fields

states::Array{Int,1}: array of states

actions::Array{Int,1}: array of actions

γ::Float64: discount factor

α::Float64: learning rate

Q::Array{Float64,2}: Q-value table

VLDataScienceMachineLearningPackage.solve — Method

function solve(model::MyQLearningModel, environment::T, startstate::Int, maxsteps::Int;
    ϵ::Float64 = 0.2) -> MyQLearningModel where T <: AbstractWorldModel

Simulate the Q-Learning agent in the given environment starting from the given state for a maximum number of steps.

Arguments

agent::MyQLearningAgentModel: The Q-Learning agent model.

environment::MyRectangularGridWorldModel: The environment model.

maxsteps::Int: The maximum number of steps to simulate.

δ::Float64 = 0.02: The convergence threshold. Default is 0.02.

worldmodel::Function = _world: The world model function. Default is the private _world function.

Returns

MyQLearningAgentModel: The updated Q-Learning agent model after simulation.