Iterative Solvers for Linear Systems

We provide several iterative solvers for linear systems, including Jacobi, Gauss-Seidel, and Successive Over-Relaxation methods. These solvers are useful for large, sparse systems where direct methods may be inefficient.

VLDataScienceMachineLearningPackage.solve — Function

solve(A::AbstractMatrix{T}, b::AbstractVector{T}, xₒ::AbstractVector{T}; 
algorithm::AbstractLinearSolverAlgorithm = JacobiMethod(), ϵ::Float64 = 0.01, maxiterations::Int64 = 100) where T <: Number

The solve function solves the linear system of equations Ax = b using the specified algorithm. The function returns the solution vector x for each iteration of an iterative method.

Arguments

A::AbstractMatrix{T}: The system matrix A in the linear system of equations Ax = b.

b::AbstractVector{T}: The right-hand side vector b in the linear system of equations Ax = b.

xₒ::AbstractVector{T}: The initial guess for the solution vector x.

algorithm::AbstractLinearSolverAlgorithm: The algorithm to use to solve the linear system of equations. The default algorithm is JacobiMethod().

ϵ::Float64: The error tolerance for the iterative method. The default value is 0.01.

maxiterations::Int64: The maximum number of iterations for the iterative method. The default value is 100.

ω::Float64: The relaxation factor for the Successive Over-Relaxation method. The default value is 0.01. This parameter is only used if the SuccessiveOverRelaxationMethod algorithm is selected.

Returns

d::Dict{Int,Array{T,1}}: The solution vector x for each iteration of an iterative method. The keys of the dictionary are the iteration numbers, and the values are the solution vectors at each iteration.