Working with Text Data

We'll work with text data in many applications. We've included a few methods to help with text processing.

VLDataScienceMachineLearningPackage.tokenize — Function

function tokenize(s::String, tokens::Dict{String, Int64}; 
    pad::Int64 = 0, padleft::Bool = false, delim::Char = ' ') -> Array{Int64,1}

Arguments

s::String - the string to tokenize.

tokens::Dict{String, Int64} - a dictionary of tokens in alphabetical order (key: token, value: position) for the entire document.

pad::Int64 - (optional) the number of padding tokens to add to the end of the tokenized string. Default is 0.

padleft::Bool - (optional) if true, the padding tokens are added to the beginning of the tokenized string. Default is false.

delim::Char - (optional) the delimiter used in the string. Default is ' '.

Returns

Array{Int64,1} - an array of integers representing the vectorized string.

VLDataScienceMachineLearningPackage.featurehashing — Function

function featurehashing(text::Array{String,1}; d::Int64 = 100, 
    algorithm::AbstractFeatureHashingAlgorithm = UnsignedFeatureHasing()) -> Array{Int64,1}

Computes the feature hashing of the input text using the specified algorithm.

Arguments

text::Array{String,1} - an array of strings to be hashed.

d::Int64 - (optional) the size of the hash table. Default is 100.

algorithm::AbstractFeatureHasingAlgorithm - (optional) the hashing algorithm to use. Default is UnsignedFeatureHasing.

Returns

Array{Int64,1} - an array of integers representing the hashed features.

function featurehashing(text::Array{Int,1}; d::Int64 = 100, 
    algorithm::AbstractFeatureHashingAlgorithm = UnsignedFeatureHasing()) -> Array{Int64,1}

Computes the feature hashing of the input text using the specified algorithm.

Arguments

text::Array{Int,1} - an array of integers to be hashed (e.g., tokenized text).

d::Int64 - (optional) the size of the hash table. Default is 100.

algorithm::AbstractFeatureHasingAlgorithm - (optional) the hashing algorithm to use. Default is UnsignedFeatureHasing.

Returns

Array{Int64,1} - an array of integers representing the hashed features.

English Language Vocabulary Model

We have included a simple vocabulary model that we can use to analyze text data. This model is based on a dictionary of common English words. We can use this model to compute a transition matrix that describes the probability of transitioning from one letter to another in a word. This can be useful for various applications, including text generation and analysis.

VLDataScienceMachineLearningPackage.vocabulary_transition_matrix — Function

vocabulary_transition_matrix(wordsdictionary::Dict{Char, Set{String}}, 
    letters::Array{Char,1}) -> Array{Float64,2}

This function computes the transition matrix for {a,b,c,...,z} given the vocabulary model and the letters.

Arguments

wordsdictionary::Dict{Char, Set{String}}: This is the vocabulary model that we have created.

letters::Array{Char,1}: This is the array of characters that we are interested in.

Returns

Array{Float64,2}: This returns back the transition matrix.

VLDataScienceMachineLearningPackage.sample_words — Function

sample_words(P::Array{Float64,2}, letters::Array{Char,1};
    number_of_samples::Int64 = 100, length_of_sample_word::Int64 = 4, startchar::Char = 'z') -> Dict{Int64,String}

This function generates sample words of a specified length from the transition matrix P given the letters.

Arguments

P::Array{Float64,2}: This is the transition matrix.

letters::Array{Char,1}: This is the array of characters that we are interested in.

number_of_samples::Int64: This is the number of samples that we want to generate.

length_of_sample_word::Int64: This is the length of the sample word that we want to generate.

startchar::Char: This is the starting character of the sample word.

Returns

Dict{Int64,String}: This returns back a dictionary of sample words. The key is the index of the sample word and the value is the sample word.