Bandit Algorithms

This section covers the implementation of several multi-armed bandit algorithms in the VLDataScienceMachineLearningPackage.jl package.

VLDataScienceMachineLearningPackage.MyExploreFirstAlgorithmModel — Type

mutable struct MyExploreFirstAlgorithmModel <: AbstractBanditAlgorithmModel

A mutable struct that represents an explore-first bandit algorithm model.

VLDataScienceMachineLearningPackage.MyEpsilonGreedyAlgorithmModel — Type

mutable struct MyEpsilonGreedyAlgorithmModel <: AbstractBanditAlgorithmModel

A mutable struct that represents an epsilon-greedy bandit algorithm model.

VLDataScienceMachineLearningPackage.MyUCB1AlgorithmModel — Type

mutable struct MyUCB1AlgorithmModel <: AbstractBanditAlgorithmModel

A mutable struct that represents a UCB1 bandit algorithm model.

VLDataScienceMachineLearningPackage.solve — Method

solve(model::AbstractBanditAlgorithmModel; T::Int = 0, world::Function = _null)

Solve the bandit problem using the given model.

Arguments

model::AbstractBanditAlgorithmModel: The model to use to solve the bandit problem.

T::Int = 0: The number of rounds to play. Default is 0.

world::Function = _null: The function that returns the reward for a given action. Default is the private _null function.

Returns

Array{Float64,2}: The rewards for each arm at each round.

VLDataScienceMachineLearningPackage.regret — Function

regret(rewards::Array{Float64,2})::Array{Float64,1}

Compute the regret for the given rewards.

Arguments

rewards::Array{Float64,2}: The rewards for each arm at each round.

Returns

Array{Float64,1}: The regret at each round.

Binary arms

Suppose we wanted to look at a bandit problem where each arm is represented by a binary vector of length K. This means that there are a total of 2^K possible arms, each corresponding to a unique combination of the K binary features. Further, the reward structure is computed in a user specified world function that maps each binary vector to a reward, along with a context model.

We've built in support for this type of bandit problem in the package. Specifically, we have defined a new abstract type AbstractBanditAlgorithmContextModel that can be used to represent the context of a bandit problem with binary arms. We have a concrete implementation of this type called MyConsumerChoiceBanditContextModel, which allows users to specify the average rewards for each possible combination of binary features.

VLDataScienceMachineLearningPackage.MyBinaryVectorArmsEpsilonGreedyAlgorithmModel — Type

mutable struct MyBinaryVectorArmsEpsilonGreedyAlgorithmModel <: AbstractBanditAlgorithmModel

A mutable struct that represents a binary vector arms epsilon-greedy bandit algorithm model.

Fields

K::Int64: number of arms

VLDataScienceMachineLearningPackage.MyConsumerChoiceBanditContextModel — Type

mutable struct MyConsumerChoiceBanditContextModel <: AbstractBanditProblemContextModel

This struct defines a consumer choice bandit context model for bandit problems with context.

Fields

data::Dict{String, Any}: A dictionary containing the context data for each item.

items::Array{String,1}: An array of item names (things we are purchasing).

bounds::Array{Float64,2}: A 2D array defining the bounds on the items that we can purchase.

B::Float64: The budget available to spend on the collection of items.

nₒ::Array{Float64,1}: An initial guess for the solution (quantities of each item).

μₒ::Array{Float64,1}: An initial estimate for the utility of each arm (collection of items).

VLDataScienceMachineLearningPackage.solve — Method

solve(model::AbstractBanditAlgorithmModel, context::AbstractBanditProblemContextModel;
    T::Int = 0, world::Function = _null)

Solve the contextual bandit problem using the given model and context.

Arguments

model::MyBinaryVectorArmsEpsilonGreedyAlgorithmModel: The model to use to solve the bandit problem.

context::MyConsumerChoiceBanditContextModel: The context model to use. Must be a subtype of AbstractBanditProblemContextModel.

T::Int = 0: The number of rounds to play. Default is 0.

world::Function = _null: The function that returns the reward for a given action. Default is the private _null function.

startdayindex::Int = 1: The starting time index for the context model. Default is 1. This is useful for time series data.