Factory methods

We use a particular pattern to build and configure the various composite types in our system. This pattern involves the use of factory functions that encapsulate the construction logic for each type, ensuring that all necessary dependencies are properly initialized and configured.

VLDataScienceMachineLearningPackage.build — Function

function build(modeltype::Type{T}, 
    data::NamedTuple)::T where T <: AbstractLinearProgrammingProblemType

The function builds a linear programming problem model from the data provided.

Arguments

modeltype::Type{T}: the type of the model to build where T is a subtype of AbstractLinearProgrammingProblemType.

data::NamedTuple: the data to use to build the model.

The data::NamedTuple must have the following fields:

A::Array{Float64,2}: the constraint matrix.

b::Array{Float64,1}: the right-hand side vector.

c::Array{Float64,1}: the cost vector.

lb::Array{Float64,1}: the lower bounds vector.

ub::Array{Float64,1}: the upper bounds vector.

Returns

a linear programming problem model of type T where T is a subtype of AbstractLinearProgrammingProblemType.

build(modeltype::Type{MyPerceptronClassificationModel}, 
    data::NamedTuple) -> MyPerceptronClassificationModel

The function builds a perceptron classification model from the data provided.

Arguments

modeltype::Type{MyPerceptronClassificationModel}: the type of the model to build.

data::NamedTuple: the data to use to build the model.

The data::NamedTuple must have the following fields:

parameters::Vector{Float64}: the coefficients of the model.

mistakes::Int64: the number of mistakes that are are willing to make.

Returns

a perceptron classification model.

build(modeltype::Type{MyLogisticRegressionClassificationModel}, 
    data::NamedTuple) -> MyLogisticRegressionClassificationModel

The function builds a logistic regression classification model from the data provided.

Arguments

modeltype::Type{MyLogisticRegressionClassificationModel}: the type of the model to build.

data::NamedTuple: the data to use to build the model.

The data::NamedTuple must have the following fields:

parameters::Vector{Float64}: the coefficients of the model.

Returns

a logistic regression classification model.

build(modeltype::Type{MyKNNClassificationModel}, 
    data::NamedTuple) -> MyKNNClassificationModel

The function builds a KNN classification model from the data provided.

Arguments

modeltype::Type{MyKNNClassificationModel}: the type of the model to build.

data::NamedTuple: the data to use to build the model.

The data::NamedTuple must have the following fields:

K::Int64: the number of neighbours to look at.

d::Function: the similarity function.

X::Matrix{Float64}: the training features.

y::Vector{Int64}: the training labels.

function build(type::Type{MyAdjacencyRecombiningCommodityPriceTree}, data::NamedTuple) -> MyAdjacencyRecombiningCommodityPriceTree

Builds an MyAdjacencyRecombiningCommodityPriceTree model given the data in the NamedTuple. This method builds the connectivity of the tree. To compute the price at each node, use the populate! method.

Arguments

type::Type{MyAdjacencyRecombiningCommodityPriceTree}: The type of the model to build.

data::NamedTuple: The data to use to build the model.

The data NamedTuple must contain the following fields:

h::Int64: The height of the tree.

price::Float64: The price at the root node.

u::Float64: The price increase factor.

d::Float64: The price decrease factor.

Returns

MyAdjacencyRecombiningCommodityPriceTree: the price tree model holding the computed price data.

function build(modeltype::Type{MyFullGeneralAdjacencyTree}, data::NamedTuple) -> MyFullGeneralAdjacencyTree

This function builds a MyFullGeneralAdjacencyTree model given the data in the NamedTuple. It populates the connectivity of the tree. However, it does not populate the data for the tree nodes. We populate the data using the populate! method.

Arguments

modeltype::Type{MyFullGeneralAdjacencyTree}: The type of the model to build.

data::NamedTuple: The data to use to build the model. The NamedTuple must have the following fields:

h::Int64: The height of the tree.

n::Int64: The branching factor of the tree.

Returns

MyFullGeneralAdjacencyTree: The constructed tree model.

function build(modeltype::Type{MyOneDimensionalElementaryWolframRuleModel}, data::NamedTuple) -> MyOneDimensionalElementarWolframRuleModel

This build method constructs an instance of the MyOneDimensionalElementaryWolframRuleModel type using the data in a NamedTuple.

Arguments

modeltype::Type{MyOneDimensionalElementaryWolframRuleModel}: The type of model to build, in this case, the MyOneDimensionalElementaryWolframRuleModel type.

data::NamedTuple: The data to use to build the model.

The data::NamedTuple must contain the following keys:

index::Int64: The index of the Wolfram rule

colors::Int64: The number of colors in the rule

radius::Int64: The radius, i.e., the number of cells to consider in the rule

Return

This function returns a populated instance of the MyOneDimensionalElementaryWolframRuleModel type.

function build(model::Type{T}, edgemodels::Dict{Int64, MyGraphEdgeModel}) where T <: AbstractGraphModel

This function builds a graph model from a dictionary of edge models.

Arguments

model::Type{T}: The type of graph model to build, where T is a subtype of AbstractGraphModel.

edgemodels::Dict{Int64, MyGraphEdgeModel}: A dictionary of edge models to use for building the graph.

Returns

T: The constructed graph model, where T is a subtype of AbstractGraphModel.

function build(modeltype::Type{MyDirectedBipartiteGraphModel}, data::NamedTuple) -> MyDirectedBipartiteGraphModel

This function builds a mutable MyDirectedBipartiteGraphModel instance given the data in the data::NamedTuple argument.

Arguments

modeltype::Type{MyDirectedBipartiteGraphModel} - The type of the model to build.

data::NamedTuple - The data to populate the model with.

The data::NamedTuple argument must contain the following fields:

s::Int64 - The source node index.

t::Int64 - The target node index.

edges::Dict{Int, MyConstrainedGraphEdgeModel} - The edges dictionary containing the constrained graph edges models.

build(model::Type{MyMDPProblemModel}, data::NamedTuple) -> MyMDPProblemModel

Builds a MyMDPProblemModel from a NamedTuple.

Arguments

model::Type{MyMDPProblemModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

𝒮::Array{Int64,1}: state space

𝒜::Array{Int64,1}: action space

T::Union{Function, Array{Float64,3}}: transition matrix of function

R::Union{Function, Array{Float64,2}}: reward matrix or function

γ::Float64: discount factor

Returns

MyMDPProblemModel: the built MDP problem model

build(modeltype::Type{MyRectangularGridWorldModel}, data::NamedTuple) -> MyRectangularGridWorldModel

Builds a MyRectangularGridWorldModel from data in a NamedTuple.

Arguments

modeltype::Type{MyRectangularGridWorldModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

nrows::Int: number of rows in the grid

ncols::Int: number of columns in the grid

rewards::Dict{Tuple{Int,Int},Float64}: dictionary of state to reward mapping

defaultreward::Float64: default reward value (optional)

Returns

MyRectangularGridWorldModel: a populated rectangular grid world model

build(type::MySimpleCobbDouglasChoiceProblem, data::NamedTuple) -> MySimpleCobbDouglasChoiceProblem

Builds a MySimpleCobbDouglasChoiceProblem model from data in a NamedTuple.

Arguments

type::MySimpleCobbDouglasChoiceProblem: the type of model to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

α::Array{Float64,1}: the Cobb-Douglas parameters

c::Array{Float64,1}: the costs of goods

initial::Array{Float64,1}: the initial endowment of goods

bounds::Array{Float64,2}: the bounds on the choice variables

I::Float64: the budget we can spend

Returns

MySimpleCobbDouglasChoiceProblem: the built Cobb-Douglas choice problem model

build(modeltype::Type{MyEpsilonGreedyAlgorithmModel}, data::NamedTuple) -> MyEpsilonGreedyAlgorithmModel

Builds a MyEpsilonGreedyAlgorithmModel from data in a NamedTuple.

Arguments

modeltype::Type{MyEpsilonGreedyAlgorithmModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

K::Int64: number of arms

build(modeltype::Type{MyExploreFirstAlgorithmModel}, data::NamedTuple) -> MyExploreFirstAlgorithmModel

Builds a MyExploreFirstAlgorithmModel from data in a NamedTuple.

Arguments

modeltype::Type{MyExploreFirstAlgorithmModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

K::Int64: number of arms

build(modeltype::Type{MyUCB1AlgorithmModel}, data::NamedTuple) -> MyUCB1AlgorithmModel

Builds a MyUCB1AlgorithmModel from data in a NamedTuple.

Arguments

modeltype::Type{MyUCB1AlgorithmModel}: the model type to

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

K::Int64: number of arms

build(modeltype::Type{MyBinaryWeightedMajorityAlgorithmModel}, 
    data::NamedTuple) -> MyBinaryWeightedMajorityAlgorithmModel

Build a Binary Weighted Majority Algorithm model. This function initializes the model with the given parameters in the data NamedTuple. The model is returned to the caller.

Arguments

modeltype::Type{MyBinaryWeightedMajorityAlgorithmModel}: the type of the model to build

data::NamedTuple: the parameters to initialize the model

The named tuple data must have the following fields:

ϵ::Float64: learning rate

n::Int64: number of experts

T::Int64: number of rounds

expert::Function: expert function

adversary::Function: adversary function

function build(modeltype::Type{MyTwoPersonZeroSumGameModel}, 
    data::NamedTuple)::MyTwoPersonZeroSumGameModel

This method builds and returns an instance of the MyTwoPersonZeroSumGameModel type.

Arguments

modeltype::Type{MyTwoPersonZeroSumGameModel}: The model type to build.

data::NamedTuple: A named tuple containing the model parameters:

ϵ::Float64: The learning rate.

n::Int: The number of experts (actions).

T::Int: The number of rounds.

payoffmatrix::Array{Float64,2}: The payoff matrix.

Returns

model::MyTwoPersonZeroSumGameModel: An instance of the MyTwoPersonZeroSumGameModel type with the specified parameters.

build(modeltype::Type{MyBinaryVectorArmsEpsilonGreedyAlgorithmModel}, data::NamedTuple) -> MyBinaryVectorArmsEpsilonGreedyAlgorithmModel

Builds a MyBinaryVectorArmsEpsilonGreedyAlgorithmModel from data in a NamedTuple.

Arguments

modeltype::Type{MyBinaryVectorArmsEpsilonGreedyAlgorithmModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

K::Int64: number of arms

Returns

MyBinaryVectorArmsEpsilonGreedyAlgorithmModel: a populated binary vector arms epsilon greedy algorithm model

build(modeltype::Type{MyConsumerChoiceBanditContextModel}, data::NamedTuple) -> MyConsumerChoiceBanditContextModel

Builds a MyConsumerChoiceBanditContextModel from data in a NamedTuple.

Arguments

modeltype::Type{MyConsumerChoiceBanditContextModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

data::Dict{String, Any}: data dictionary for each item, or more generally the context

items::Array{String,1}: items for each asset

bounds::Array{Float64,2}: bounds on the assets that we can purchase

B::Float64: budget that we have to spend on the collection of assets

nₒ::Array{Float64,1}: initial guess for the solution

μₒ::Array{Float64,1}: initial for the utility of each arm

γ::Array{Float64,1}: parameters for the utility function (preferences)

Returns

MyConsumerChoiceBanditContextModel: a populated consumer choice bandit context model

function build(type::Type{MyQLearningModel},data::NamedTuple) -> MyQLearningModel

Builds a MyQLearningAgentModel from data in a NamedTuple.

Arguments

modeltype::Type{MyQLearningAgentModel}: the model type to build

data::NamedTuple: the data to use to build the model

The data NamedTuple must contain the following keys:

states::Array{Int64,1}: the state space

actions::Array{Int64,1}: the action space

α::Float64: the learning rate

γ::Float64: the discount factor

Returns

MyQLearningAgentModel: a populated Q-learning agent model

build(modeltype::Type{MyClassicalHopfieldNetworkModel}, data::NamedTuple) -> MyClassicalHopfieldNetworkModel

Factory method for building a Hopfield network model.

Arguments

modeltype::Type{MyClassicalHopfieldNetworkModel}: the type of the model to be built.

data::NamedTuple: a named tuple containing the data for the model.

The named tuple should contain the following fields:

memories: a matrix of memories (each column is a memory).

Returns

model::MyClassicalHopfieldNetworkModel: the built Hopfield network model with the following fields populated:

W: the weight matrix.

b: the bias vector.

energy: a dictionary of energies for each memory.

build(modeltype::Type{MyModernHopfieldNetworkModel}, data::NamedTuple) -> MyModernHopfieldNetworkModel

Factory method for assembling a modern Hopfield network model from raw memories and an inverse-temperature parameter.

Arguments

modeltype::Type{MyModernHopfieldNetworkModel}: concrete model type to instantiate.

data::NamedTuple: expects memories (matrix with memories on columns) and β (inverse-temperature scalar).

Returns

model::MyModernHopfieldNetworkModel: model populated with X (memory matrix) and β.

build(modeltype::Type{MySimpleBoltzmannMachineModel}, data::NamedTuple) -> MySimpleBoltzmannMachineModel

Factory method for building a simple Boltzmann machine model.

Arguments

modeltype::Type{MySimpleBoltzmannMachineModel}: concrete model type to instantiate.

data::NamedTuple: expects W::Array{Float64, 2} (weight matrix) and b::Vector{Float64} (bias vector).

Returns

model::MySimpleBoltzmannMachineModel: model populated with W and b.

build(modeltype::Type{MyRestrictedBoltzmannMachineModel}, data::NamedTuple) -> MyRestrictedBoltzmannMachineModel

Factory method for building a restricted Boltzmann machine model.

Arguments

modeltype::Type{MyRestrictedBoltzmannMachineModel}: concrete model type to instantiate.

data::NamedTuple: the data to use to build the model. The NamedTuple must have the following fields:

W::Array{Float64,2}: weight matrix between visible and hidden units.

b::Vector{Float64}: bias vector for visible units.

a::Vector{Float64}: bias vector for hidden units.

Returns

model::MyRestrictedBoltzmannMachineModel: model populated with W, b, and a.

build(modeltype::Type{MyNaiveKMeansClusteringAlgorithm}, data::NamedTuple) -> MyNaiveKMeansClusteringAlgorithm

Factory method for building a naive K-Means clustering algorithm model.

Arguments

modeltype::Type{MyNaiveKMeansClusteringAlgorithm}: concrete model type to instantiate.

data::NamedTuple: expects the following fields:

K::Int64: number of clusters

ϵ::Float64: convergence threshold

maxiter::Int64: maximum number of iterations

dimension::Int64: dimensionality of the data points

number_of_points::Int64: number of data points

scale_factor::Float64: scale factor for initializing centroids

Returns

model::MyNaiveKMeansClusteringAlgorithm: model populated with initial assignments and centroids.

VLDataScienceMachineLearningPackage.populate! — Function

populate!(model::MyAdjacencyRecombiningCommodityPriceTree, price::Float64, Δ::Array{Float64,1})

This function populates the price tree model with the given price and price change factors. This method updates the model in place.

Arguments

model::MyAdjacencyRecombiningCommodityPriceTree: The price tree model to populate.

price::Float64: The initial price to set at the root of the tree.

Δ::Array{Float64,1}: The array of price change factors for each level of the tree.

function populate!(model::MyFullGeneralAdjacencyTree, configuration::Function)::MyFullGeneralAdjacencyTree

Populates the data for the tree model using the provided configuration function.

Arguments

model::MyFullGeneralAdjacencyTree: The tree model instance to populate.

configuration::Function: A function that takes four arguments (level, index, offset and parentdatapayload) and returns a NamedTuple with the configuration data for that node.

Returns

MyFullGeneralAdjacencyTree: The updated tree model with populated data.