Linear Systems

We provide several techniques for working with linear systems, including eigendecomposition using QR iteration and iterative solvers for linear systems. These solvers are useful for large, sparse systems where direct methods may be inefficient.

In addition, we provide functionality for solving linear programming problems.

VLDataScienceMachineLearningPackage.qriteration — Function

qriteration(A::Array{Float64,2}; maxiter::Int64 = 10, tolerance::Float64 = 1e-9) -> Tuple

Computes the eigenvalues and eigenvectors of a real matrix A using the QR iteration method.

Arguments

A::Array{Float64,2}: a real matrix of size n x n.

maxiter::Int64: the maximum number of iterations (default is 10).

tolerance::Float64: the tolerance for the stopping criterion (default is 1e-9).

Returns

Tuple: a tuple of two elements: the first element is an array of eigenvalues and the second element is a dictionary of eigenvectors.

VLDataScienceMachineLearningPackage.solve — Function

solve(A::AbstractMatrix{T}, b::AbstractVector{T}, xₒ::AbstractVector{T}; 
algorithm::AbstractLinearSolverAlgorithm = JacobiMethod(), ϵ::Float64 = 0.01, maxiterations::Int64 = 100) where T <: Number

The solve function solves the linear system of equations Ax = b using the specified algorithm. The function returns the solution vector x for each iteration of an iterative method.

Arguments

A::AbstractMatrix{T}: The system matrix A in the linear system of equations Ax = b.

b::AbstractVector{T}: The right-hand side vector b in the linear system of equations Ax = b.

xₒ::AbstractVector{T}: The initial guess for the solution vector x.

algorithm::AbstractLinearSolverAlgorithm: The algorithm to use to solve the linear system of equations. The default algorithm is JacobiMethod().

ϵ::Float64: The error tolerance for the iterative method. The default value is 1e-6.

maxiterations::Int64: The maximum number of iterations for the iterative method. The default value is 1000.

ω::Float64: The relaxation factor for the Successive Over-Relaxation method. The default value is 1.0. This parameter is only used if the SuccessiveOverRelaxationMethod algorithm is selected.

Returns

d::Dict{Int,Array{T,1}}: The solution vector x for each iteration of an iterative method. The keys of the dictionary are the iteration numbers, and the values are the solution vectors at each iteration.

solve(problem::MyLinearProgrammingProblemModel) -> Dict{String,Any}

Solves a linear programming problem defined by the MyLinearProgrammingProblemModel instance using the GLPK solver.

Arguments

problem::MyLinearProgrammingProblemModel: An instance of MyLinearProgrammingProblemModel holding the data for the problem.

constraints::Symbol: The type of constraints to apply. Options are :leq (less than or equal to), :geq (greater than or equal to), or :eq (equal to). Default is :leq.

Returns

Dict{String,Any}: A dictionary with the following keys:

"argmax": The optimal choice.

"budget": The budget at the optimal choice.

"objective_value": The value of the objective function at the optimal choice.